Added CPU & MEM graphs

+++ src-tauri/src/our_mods/proc_functions.rs
@@ -45,7 +45,7 @@ pub fn update_procs(pid_table: &mut HashMap<u32, u16>, procs: &mut Vec<Process>,
         let idled = idle as f64 - sys_stats._idle as f64;
 
         // ((stat.utime+stat.stime)/ticks_per_second()) as f32  / (uptime as f32 - (stat.starttime/ticks_per_second()) as f32)
-        cpus_usage.push( 1- (idled) as f32/ totald as f32);
+        cpus_usage.push( (1 as f32 - idled as f32) as f32/ totald as f32);
         //cpus_usage.push( (totald - idled) as f32/ totald as f32);
         
         sys_stats._idle = idle;
